# Conflicts:
#	.docker/rootfs/etc/php7/conf.d/99-tasmoadmin.ini


Former-commit-id: 8378d4dea080df81585f6cb662f6e505f185ecca

+++ tasmoadmin/pages/upload.php
@@ -7,7 +7,7 @@
 
 	$files = glob( $firmwarefolder.'*' ); // get all file names
 	foreach( $files as $file ) { // iterate files
-		if( is_file( $file ) && strpos( $file, ".empty" ) === FALSE ) {
+		if( is_file( $file ) && strpos( $file, ".empty" ) === FALSE && strpos( $file, ".htaccess" ) === FALSE ) {
 			unlink( $file );
 		} // delete file
 	}
@@ -234,9 +234,14 @@
 	}
 
 
-	$ota_server_ip = isset( $_REQUEST[ "ota_server_ip" ] ) ? $_REQUEST[ "ota_server_ip" ] : "";
+	$ota_server_ssl  = isset( $_REQUEST[ "ota_server_ssl" ] ) ? $_REQUEST[ "ota_server_ssl" ] : "0";
+	$ota_server_ip   = isset( $_REQUEST[ "ota_server_ip" ] ) ? $_REQUEST[ "ota_server_ip" ] : "";
+	$ota_server_port = isset( $_REQUEST[ "ota_server_port" ] ) ? $_REQUEST[ "ota_server_port" ] : "";
 
+
+	$Config->write( "ota_server_ssl", $ota_server_ssl );
 	$Config->write( "ota_server_ip", $ota_server_ip );
+	$Config->write( "ota_server_port", $ota_server_port );
 
 ?>
 
@@ -730,7 +735,7 @@
 
 		} );
 	</script>
-	<script type='text/javascript'
-	        src='<?php echo _RESOURCESURL_; ?>js/devices.js?<?php echo time(); ?>'></script>
+
+	<script src="<?php echo URL::JS( "devices" ); ?>"></script>
 <?php endif; ?>
